fix: solve repeated merge

+++ triedb/pathdb/async_incr_state_test.go
@@ -249,7 +249,7 @@ func TestFlush(t *testing.T) {
 		if len(entry.Nodes) > 0 {
 			jn = append(jn, entry)
 			entrySize := ownerSize + nodesListSize
-			totalSize += uint64(entrySize)
+			totalSize += entrySize
 		}
 
 		if totalSize >= batchSize {
